+++ sort.h
@@ -4,10 +4,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void bubblesort(int *v, int n, int max_value);
-void mergesort(int *v, int n, int max_value);
-void metodocaixas(int *vet, int n, int max_value);
-void insertionsort(int *vet, int n, int max_value);
-void quicksort(int *v, int n, int max_value);
+void bubblesort(unsigned long *v, int n, unsigned long max_value);
+void mergesort(unsigned long *v, int n, unsigned long max_value);
+void metodocaixas(unsigned long *vet, int n, unsigned long max_value);
+void insertionsort(unsigned long *vet, int n, unsigned long max_value);
+void quicksort(unsigned long *v, int n, unsigned long max_value);
 
 #endif // SORT_H_INCLUDED
